requirements.txt
flask==3.1.0
flask-cors==4.0.0
gunicorn==21.2.0
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
